backend/requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-filter==24.3
markdown==3.7
django-cors-headers==4.4.0
gunicorn==23.0.0